Update tempest test files to fix various test issues
Issues: Fixes #20

Problem:
The scenario tests for the mitaka branch fail due to request calls to VIP not properly configuring (or bypassing) ssl verification. We need to fix up the requests to have a context that allows the request to be verified.

Analysis:
Modified the requirements.txt file for mitaka to point to a tempest release that is known to work (12.1.0). Updated the request call to the VIP to have an ssl context.

Tests:
Ran tests manually in boulder lab.