Test session persistence scenario test in fails in mitaka
Issues:
Fixes #32

Problem:
Once session persistence has been set on a pool, it is impossible,
through the CLI, to set it back to None. It is possible to set it to
None via the api, but the test has a bug where it is not set properly.
We can fixit by calling the rest endpoint with the session_persistence
key set in the pool dict and the value equal to None.

Analysis:
Modified the base.py module for the scenario tests to set
session_persistence to None in the pool update_data dict when the
session persistence is updated for a given pool. This allows the
session_persistence to be set back to None.

Tests:
Ran test_session_persistence scenario tests against mitaka stack.