stable/newton api test timeout is set at 10 minutes
Issues:
Fixes #34

Problem:
In the tempest api tests, the timeout for waiting for a loadbalancer to
become active and online is 600 seconds. We have brought this time down
to 60 seconds for other openstack releases. We should do the same for
newton.

Analysis:
Changed the timeout to 60 seconds.

Tests:
Ran api tests on a local stable/newton stack.