stable/newton pool member api tests need to use IP address other than loopback
Issues:
Fixes #39

Problem:
There are a few member api tests which use the loopback address
127.0.0.1 to create dummy members, but the BIG-IP does not allow members
to be created with this address, so those tests fails. We made this
change in liberty and mitaka, but since we are not merging up all those
changes into newton, we need to cherry-pick those commits into the
stable/newton branch.

Analysis:
Changed 127 adresses to 128 address where necessary.

Tests:
Ran the driver tests, which include the tests here, against a newton
stack.