Teardown robustness needed in tempest tests
Issues:
Fixes #43

Problem:
Our tempest tests are not immune to test interactions. We need to harden
the test teardown process as much as we can currently, allowing tests to
properly teardown resources created by the test through the proper
channels (neutron REST api). This just means making the cleanup calls
for the tests and test suites be more flexible in the way it detects and
moves around errors.

Analysis:
Moved teardown logic for api tests into the neutron-lbaas repo, so
the driver can benefit from the code as well. This means that two PRs
will be in flight for this change, one to move the logic over to this
repo, and one to remove the logic in the driver.

Tests:
Ran tests against my local fork. Saw an intermittent Bad Status Line
failure in the test_listener_basic scenario test, but I haven't nailed
down the cause, other than bad test assumptions. Will keep an eye on
this in nightly.